+++ tests/Fake/FooFoo.php
@@ -1,0 +1,8 @@
+<?php
+declare(strict_types=1);
+
+namespace Capsule\Di\Fake;
+
+class FooFoo extends Foo
+{
+}