+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## NEXT
+
+- Better handling of optional arguments
+
+- By default, child classes now inherit the constructor arguments of their
+  parent class. This means you do not need to re-define the child class
+  arguments if you want them to be the same as the parent.You can still
+  override arguments in the child class definition. Does not apply to property
+  injection or setters, or to factory, or to class() overrides. To turn off
+  inherited arguments for a class, call $def->{Foo::CLASS}->inherit(null).
+
 ## 3.3.0
 
 - Add ArrayValues::merge()